Replaces jQuery-UJS with non-jQuery alternative, drops jQuery altogether
Bootstrap 5 doesn't require jQuery anymore and UJS was the only holdout.
Many frontend devs agree that the browser APIs have sufficiently
stabilized and centered on a user-friendly standard interface such that
jQuery is no longer needed. See https://youmightnotneedjquery.com/ for
more.

Given Ruby on Rails has dropped jQuery as a dependency since 5.x
(current is 7.x), it seems it's about time Buffalo does the same.

rails-ujs is a nearly drop-in replacement for jquery-ujs. For
scaffolding purposes, it is a minor change in a couple files (this
commit alone) and no changes to how forms need to be rendered. Also of
note is that rails-ujs can still work with jQuery if desired.