build: Allow custom module import names
The existing code uses gobuffalo/meta to determine the module name, but
strips the fully qualified name from it. This is fine when dealing with
a buffalo project at the top level of a git repository/go module. However, if
multiple projects share a single go.mod file higher up the directory
tree this detection fails.

Using a build-time flag, restore the original fully qualified name of
the module so the buffalo project can be built as a sub-package. Further
ensure that actions/models/grifts are included in this.

An alternative solution would be to implement this as config option in
gobuffalo/meta. I believe this is the cleaner solution, however, given
https://github.com/gobuffalo/meta/issues/11 the current state of meta
seems in question. I've implemented this directly in gobuffalo/cli
instead.